{{cookiecutter.project_slug}}/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
backcall==0.1.0           # via ipython
click-help-colors==0.4
click==6.7
colorlog==3.1.4           # via seveno-pyutil
decorator==4.3.0          # via ipython, traitlets
ipython-genutils==0.2.0   # via traitlets
ipython==6.4.0
jedi==0.12.1              # via ipython
marshmallow==2.15.3       # via seveno-pyutil
objproxies==0.9.4
parso==0.3.0              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.6.0         # via pexpect
pygments==2.2.0           # via ipython, seveno-pyutil
python-prctl==1.7
pytz==2018.5              # via seveno-pyutil, tzlocal
pyyaml==3.13
seveno-pyutil==0.4.2
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via prompt-toolkit, traitlets
sqlparse==0.2.4           # via seveno-pyutil
traitlets==4.3.2          # via ipython
tzlocal==1.5.1            # via seveno-pyutil
wcwidth==0.1.7            # via prompt-toolkit
